# WebAPIUsingOpenAPI/VectoryPathwayAPI/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "description" : "This is a sample server example Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "Swagger Petstore",
    "version" : "1.0.7"
  },
...
  "servers" : [ {
    "url" : "https://petstore.swagger.io/v2"
  }, {
    "url" : "http://petstore.swagger.io/v2"
  } ],
...
  "paths" : {
...
    "/user/createWithList" : {
      "post" : {
        "operationId" : "createUsersWithListInput",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/User"
                },
                "type" : "array"
              }
            }
          },
          "description" : "List of user object",
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : { },
            "description" : "successful operation"
          }
        },
        "summary" : "Creates list of users with given input array",
        "tags" : [ "user" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/user/{username}" : {
      "delete" : {
        "description" : "This can only be done by the logged in user.",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "description" : "The name that needs to be deleted",
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "content" : { },
            "description" : "Invalid username supplied"
          },
          "404" : {
            "content" : { },
            "description" : "User not found"
          }
        },
        "summary" : "Delete user",
        "tags" : [ "user" ]
      },
      "get" : {
        "operationId" : "getUserByName",
        "parameters" : [ {
          "description" : "The name that needs to be fetched. Use user1 for testing. ",
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : { },
            "description" : "Invalid username supplied"
          },
          "404" : {
            "content" : { },
            "description" : "User not found"
          }
        },
        "summary" : "Get user by user name",
        "tags" : [ "user" ]
      },
      "put" : {
        "description" : "This can only be done by the logged in user.",
        "operationId" : "updateUser",
        "parameters" : [ {
          "description" : "name that need to be updated",
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "description" : "Updated user object",
          "required" : true
        },
        "responses" : {
          "400" : {
            "content" : { },
            "description" : "Invalid user supplied"
          },
          "404" : {
            "content" : { },
            "description" : "User not found"
          }
        },
        "summary" : "Updated user",
        "tags" : [ "user" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/user/login" : {
      "get" : {
        "operationId" : "loginUser",
        "parameters" : [ {
          "description" : "The user name for login",
          "in" : "query",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The password for login in clear text",
          "in" : "query",
          "name" : "password",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "successful operation",
            "headers" : {
              "X-Rate-Limit" : {
                "description" : "calls per hour allowed by the user",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              },
              "X-Expires-After" : {
                "description" : "date in UTC when token expires",
                "schema" : {
                  "format" : "date-time",
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "content" : { },
            "description" : "Invalid username/password supplied"
          }
        },
        "summary" : "Logs user into the system",
        "tags" : [ "user" ]
      }
    },
...
    "/user/createWithArray" : {
      "post" : {
        "operationId" : "createUsersWithArrayInput",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/User"
                },
                "type" : "array"
              }
            }
          },
          "description" : "List of user object",
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : { },
            "description" : "successful operation"
          }
        },
        "summary" : "Creates list of users with given input array",
        "tags" : [ "user" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/user" : {
      "post" : {
        "description" : "This can only be done by the logged in user.",
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "description" : "Created user object",
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : { },
            "description" : "successful operation"
          }
        },
        "summary" : "Create user",
        "tags" : [ "user" ],
        "x-codegen-request-body-name" : "body"
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "User" : {
        "example" : {
          "firstName" : "firstName",
          "lastName" : "lastName",
          "password" : "password",
          "userStatus" : 6,
          "phone" : "phone",
          "id" : 0,
          "email" : "email",
          "username" : "username"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "username" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "userStatus" : {
            "description" : "User Status",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "User"
        }
      },
...
    },
    "securitySchemes" : {
      "api_key" : {
        "in" : "header",
        "name" : "api_key",
        "type" : "apiKey"
      },
      "petstore_auth" : {
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://petstore.swagger.io/oauth/authorize",
            "scopes" : {
              "read:pets" : "read your pets",
              "write:pets" : "modify pets in your account"
            }
          }
        },
        "type" : "oauth2"
      }
    }
  },
...
}